requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==2.4.1
Django==1.8.2
django-redis==4.10.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.7
itsdangerous==1.1.0
kombu==4.2.1
mutagen==1.41.1
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
redis==3.0.1
requests==2.20.1
rerequests==1.0.0b0
six==1.11.0
urllib3==1.24.1
vine==1.1.4
